Work around symbols breaking in complex, multi-library setups (#12)
When dragging in a symbol instance that includes a foreign symbol
(i.e. from another library), don't just call addForeignSymbol,
since that doesn't bring in additional symbols that it references.
Instead, import it from the nested library directly (if that library
is available). If not, then keep doing addForeignSymbol.

A true fix will need to either import the symbol from the top-level
library (not ideal), or add *all* foreign symbols related to the
nested one, not just the nested one.

Total inception.